+++ examples/idas/parallel/idasBruss_ASAp_kry_bbd_p.c
@@ -130,7 +130,7 @@ typedef struct {
   realtype rhs[NUM_SPECIES];
   MPI_Comm comm;
   realtype rates[2];
-  int n_local;
+  long int n_local;
 } *UserData;
 
 /* Prototypes for functions called by the IDA Solver. */
@@ -138,11 +138,11 @@ static int res(realtype tt,
                N_Vector uv, N_Vector uvp, N_Vector rr, 
                void *user_data);
 
-static int reslocal(int Nlocal, realtype tt, 
+static int reslocal(long int Nlocal, realtype tt, 
                     N_Vector uv, N_Vector uvp, N_Vector res, 
                     void *user_data);
 
-static int rescomm(int Nlocal, realtype tt,
+static int rescomm(long int Nlocal, realtype tt,
                    N_Vector uv, N_Vector uvp, 
                    void *user_data);
 
@@ -168,7 +168,7 @@ static int resB(realtype tt,
                 N_Vector yyB, N_Vector ypB, N_Vector rrB,
                 void *user_dataB);
 
-static int resBlocal(int Nlocal, realtype tt, 
+static int resBlocal(long int Nlocal, realtype tt, 
                      N_Vector uv, N_Vector uvp, 
                      N_Vector yyB, N_Vector ypB, 
                      N_Vector res, void *user_dataB);
@@ -185,8 +185,8 @@ static void SetInitialProfilesB(N_Vector uv, N_Vector uvp,
                                 N_Vector residB, UserData data);
 
 static void PrintHeader(int SystemSize, int maxl, 
-                        int mudq, int mldq, 
-                        int mukeep, int mlkeep,
+                        long int mudq, long int mldq, 
+                        long int mukeep, long int mlkeep,
                         realtype rtol, realtype atol);
 
 static void PrintOutput(void *mem, N_Vector uv, realtype time,
@@ -212,7 +212,7 @@ int main(int argc, char *argv[])
   MPI_Comm comm;
   void *mem;
   UserData data;
-  int SystemSize, local_N, mudq, mldq, mukeep, mlkeep;
+  long int SystemSize, local_N, mudq, mldq, mukeep, mlkeep;
   realtype rtol, atol, t0, tout, tret;
   N_Vector uv, uvp, resid, id, uvB, uvpB, residB, qB;
   int thispe, npes, maxl, retval;
@@ -615,8 +615,8 @@ static void SetInitialProfilesB(N_Vector uv, N_Vector uvp,
  */
 
 static void PrintHeader(int SystemSize, int maxl, 
-                        int mudq, int mldq, 
-                        int mukeep, int mlkeep,
+                        long int mudq, long int mldq, 
+                        long int mukeep, long int mlkeep,
                         realtype rtol, realtype atol)
 {
   printf("\n BRUSSELATOR: chemically reacting system\n\n");
@@ -908,7 +908,7 @@ static int res(realtype tt,
 {
   int retval;
   UserData data;
-  int Nlocal;
+  long int Nlocal;
   
   data = (UserData) user_data;
   
@@ -933,7 +933,7 @@ static int res(realtype tt,
  * and receive-waiting, in routines BRecvPost, BSend, BRecvWait.         
  */
 
-static int rescomm(int Nlocal, realtype tt, 
+static int rescomm(long int Nlocal, realtype tt, 
                    N_Vector uv, N_Vector uvp,
                    void *user_data)
 {
@@ -1149,7 +1149,7 @@ static void BSend(MPI_Comm comm, int my_pe, int ixsub, int jysub,
  * for use by the preconditioner setup routine.                          
  */
 
-static int reslocal(int Nlocal, realtype tt, 
+static int reslocal(long int Nlocal, realtype tt, 
                     N_Vector uv, N_Vector uvp, N_Vector rr,
                     void *user_data)
 {
@@ -1316,7 +1316,7 @@ static int resB(realtype tt, N_Vector yy, N_Vector yp,
 {
   UserData data;
   int retval;
-  int Nlocal;
+  long int Nlocal;
   
   data = (UserData) user_dataB;
 
@@ -1333,7 +1333,7 @@ static int resB(realtype tt, N_Vector yy, N_Vector yp,
 
 
 
-static int resBlocal(int Nlocal, realtype tt, 
+static int resBlocal(long int Nlocal, realtype tt, 
                      N_Vector uv, N_Vector uvp,
                      N_Vector uvB, N_Vector uvpB, N_Vector rrB,
                      void *user_dataB)
